+++ source/modulo_components/test/cpp/test_component.cpp
@@ -52,7 +52,7 @@ TEST_F(ComponentTest, RatePeriodParameters) {
 
 TEST_F(ComponentTest, AddRemoveOutput) {
   std::shared_ptr<State> data = make_shared_state(CartesianState::Random("test"));
-  component_->add_output("_tEsT_#1@3", data);
+  component_->add_output("8_tEsT_#1@3", data);
   EXPECT_TRUE(component_->outputs_.find("test_13") != component_->outputs_.end());
   EXPECT_NO_THROW(component_->outputs_.at("test_13")->publish());
   EXPECT_THROW(component_->publish_output("test_13"), exceptions::ComponentException);
@@ -65,9 +65,9 @@ TEST_F(ComponentTest, AddRemoveOutput) {
   component_->remove_output("test_13");
   EXPECT_TRUE(component_->outputs_.find("test_13") == component_->outputs_.end());
 
-  component_->add_output("_tEsT_#1@3", data, "", true, false);
+  component_->add_output("8_tEsT_#1@3", data, "", true, false);
   EXPECT_FALSE(component_->periodic_outputs_.at("test_13"));
-  EXPECT_NO_THROW(component_->publish_output("_tEsT_#1@3"));
+  EXPECT_NO_THROW(component_->publish_output("8_tEsT_#1@3"));
   EXPECT_NO_THROW(component_->publish_output("test_13"));
   EXPECT_THROW(component_->publish_output(""), exceptions::ComponentException);
 }