+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Release Versions:
 
 - chore: fix links in readme (#68)
 - feat: add mutex around step callback (#67)
+- feat: add warn log when topics are sanitized (#66)
 
 ## 4.0.0